Discover services, not apps
By silly mistake, the driver returns CCP application names instead of
OpenStack services in discovery info. The names are very similar to
application names and most of the time are identical. However, it's not
the case for Keystone or Horizon, for instance.

This patch fixes the bug, so now discovery info contains proper service
names.